* 'SC_#8203_Stat_Boost_Modules' of https://github.com/msborromeo/TheDoomChild:
  [*] Implemented overrides for updating Crit Chance and Damage througth modules
  [*] Added Critical Hit Handle logic to Basic Slashes and Attack Behaviour
  [*] Updated modifier value to float
  [*] Updated Modifier Value to float [*] Implemented additions to IEquipmentHandle
  [+] Added Crit Chance and Crit Damage Modules for Equipment System
  [*] Renamed MaxHPModifier script [*] Moved CriticalHitVFXHandle Script [+] Created Crit Chance Modifier Equipment Module (WIP) [+] Created Max SP Modifer Equipment Module [+] Created Resistance Modifier Equipment Module [+] Created SP Regen Rate Modifier Equipment Module

+++ Assets/DChild/Scripts/Gameplay/System/PlayerSystems/EquipmentSystem/IEquipmentStatBoostModule.cs
@@ -8,7 +8,7 @@ namespace DChild.Gameplay.EquipmentSystem
     public interface IEquipmentStatBoostModule
     {
         StatBoostType GetBoostType();
-        int GetModifierValue();
+        float GetModifierValue();
 
         void AttachTo(IPlayer player);
         void DetachFrom(IPlayer player);
